microapp/requirements.txt
psycopg2==2.8.5
requests==2.27.1
pyproj==2.6.1.post1
pandas==1.4.2
pytz==2020.1
pyqt==5.9.2
